Fix the issue of graph showing nothing
The graph in the statistics page did not show data when the page is rendered in the first place.
The graph of the current week showed data correctly only after trying to get the prev graph and going back to the current week.

what was wrong:
setWeek(week) after modifying the week state variable did not force
re-rendering due to the same reference. Thus, I cloned the week before
modifying it.